requirements.txt
depthai==2.21.2.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pytesseract==0.3.10
PyYAML==6.0
pyzbar==0.1.9
